% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modisa.R
\name{MODISA_download}
\alias{MODISA_download}
\title{Download one or more MODISA files from the NERACOOS fileServer.}
\usage{
MODISA_download(x, dest = ".")
}
\arguments{
\item{x}{one or more "DatasetsRefClass" class objects - see MODISA_query}

\item{dest}{fully qualified destination directory, if it doesn't exist an
attempt is made to create it.}
}
\value{
logical vector the same length as x.  TRUE means success
}
\description{
Download one or more MODISA files from the NERACOOS fileServer.
}

