% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neracooscrawler-package.R
\docType{package}
\name{neracooscrawler}
\alias{neracooscrawler}
\alias{neracooscrawler-package}
\title{neracooscrawler - programmatic access to Ocean Color OpeNDAP servcies.}
\description{
A crawler for programmatically working with the OpeNDAP data offerings from
}
\examples{
\dontrun{ Top <- threddscrawler::get_catalog(platform = 'MODISA', verbose = TRUE) }
}

